Update url slug logic
Add url slug to category entity and create inheritance to a base class called WebEntity because url slug is equaly determined for all entities. Changed logic from id in database to a random guid to make sure the slug is unique. This because the id is not known by the entity when first created.